fix: changes according PR review
