+++ app/src/main/res/values/main_manu_tab_ids.xml
