+++ downloads/src/main/AndroidManifest.xml
